Fixing hex to be easy to edit
I realized that my last version of the svg saved wrong - it was just PNG embedded in an SVG.  This is a proper SVG.  You can change the background color via the `fill:` attribute on line 29, or in Inkscape.

Note that the text and logo in the hex are still based on the PNG in this repository.  If @karthik has a vector version I'm happy to use that.